requirements.txt
flask==2.3.3
python-dotenv==1.0.0
redis==4.6.0